device_emulator/requirements.txt
paho-mqtt==1.6.1
python-dateutil==2.8.2
Flask==3.0.0
Flask-CORS==4.0.0
